requirements.txt
aiofiles==24.1.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
priority==2.0.0
python-dotenv==1.0.1
Quart==0.19.6
sniffio==1.3.1
Werkzeug==3.0.3
wsproto==1.2.0
